python_proj/requirements.txt
websocket-client==1.8.0
pydub==0.25.1
requests==2.32.3
pillow==11.2.1
tqdm==4.67.1